+++ src/pages/AdminPage/pages/Products/components/Products.jsx
@@ -1,14 +1,18 @@
-// Product.jsx
-import React, { useState, useEffect } from 'react';
+import React, { useState, useEffect} from 'react';
 import Sidenav from '../../../components/Sidenav';
 import Box from '@mui/material/Box';
 import FormComponent from './FormComponent';
 import axios from 'axios';
+import { useSelector } from 'react-redux';
 
 const Products = () => {
   const [showForm, setShowForm] = useState(false);
   const [products, setProducts] = useState([]);
-  const [selectedProduct, setSelectedProduct] = useState(null); 
+  const [selectedProduct, setSelectedProduct] = useState(null);
+  const [deleteMessage, setDeleteMessage] = useState('');
+  const user = useSelector(state => state.user.user);
+  const token = user ? user.token : null;
+
 
   useEffect(() => {
     fetchProducts();
@@ -35,18 +39,34 @@ const Products = () => {
 
   const handleDelete = async (productId) => {
     try {
-      const token = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjY1ZjAyOGI5NzM2ZWZkZGQ5NTZkNzVjNyIsInVzZXJuYW1lIjoibnN0X2FkbWluIiwiaWF0IjoxNzEwMjQzMDk5LCJleHAiOjE3MTAzMjk0OTl9.LqI9kLTYhfSjHtlZsW8fgoqd8VPYyneFlxj-OME5ncg"; // Replace "YOUR_AUTHENTICATION_TOKEN" with the actual token
+      // Ensure token is defined before using it
+      if (!token) {
+        console.error('Token is not available');
+        return;
+      }
+      
+      // Use the token for authorization
       await axios.delete(`https://nst-website-api.onrender.com/api/v1/products/${productId}`, {
         headers: {
           'Authorization': `Bearer ${token}`
         }
       });
-      fetchProducts(); // Refresh products after deletion
+      
+      // Refresh products after deletion
+      fetchProducts();
+  
+      // Set delete message
+      setDeleteMessage('Product deleted successfully');
+  
+      // Hide the delete message after 2 seconds
+      setTimeout(() => {
+        setDeleteMessage('');
+      }, 2000);
     } catch (error) {
       console.error('Error deleting product:', error);
     }
   };
-
+  
   const handleUpdate = () => {
     fetchProducts(); // Reload products after successful update
   };
@@ -72,13 +92,19 @@ const Products = () => {
                 <tr key={product._id}>
                   <td style={{ border: '1px solid white', padding: '8px', textAlign: 'center' }}>{product.title}</td>
                   <td style={{ border: '1px solid white', padding: '8px', textAlign: 'center' }}>
-                    <button onClick={() => handleEdit(product)}>Edit</button>
-                    <button onClick={() => handleDelete(product._id)}>Delete</button> 
+                    <button onClick={() => handleEdit(product)} style={{ marginRight: '30px' }}>Edit</button>
+                    <button onClick={() => handleDelete(product._id)}>Delete</button>
                   </td>
                 </tr>
               ))}
             </tbody>
           </table>
+          {/* Delete Message Banner */}
+          {deleteMessage && (
+            <div style={{ position: 'fixed', top: 10, right: 10, backgroundColor: 'green', color: '#fff', padding: '10px', borderRadius: '5px', zIndex: 999 }}>
+              {deleteMessage}
+            </div>
+          )}
         </Box>
       </Box>
     </>
